+++ Frameworks/MATabView/MATabView/MATabBarItem.swift
@@ -2,48 +2,28 @@
 //  MATabBarItem.swift
 //  MATabView
 //
-//  Created by Ashwin Paudel on 2022-02-02.
+//  Created by Ashwin Paudel on 2022-01-06.
 //  Copyright © 2021-2022 Ashwin Paudel. All rights reserved.
 //
-// Some code used from: https://github.com/robin/LYTabView
 
 import Cocoa
 
 @objc public protocol MATabBarItemDelegate: NSObjectProtocol {
-    @objc optional func tabBarViewItem(_ tabBarViewItem: MATabBarItem, wantsToClose tabBarViewItemIndex: Int)
+    @objc optional func tabBarItem(_ tabBarItem: MATabBarItem, wantsToClose tab: MATab)
 }
 
 open class MATabBarItem: NSButton {
-    open var item: MATabViewItem?
+    open var tab: MATab
+    var isSelectedTab = false
     open weak var delegate: MATabBarItemDelegate?
+    open var configuration = MATabViewConfiguration()
 
     let tabTitle = NSTextField(frame: .zero)
     open var closeButton = NSButton(frame: .zero)
 
-    var isMainButton = false
-
     var xPosition: CGFloat = 4
     var yPosition: CGFloat = 2
     var closeButtonSize = NSSize(width: 16, height: 16)
-    private var _favicon: NSImage?
-
-    public var tabBarView: MATabBarView?
-
-    // Drag and Droping
-    private var dragOffset: CGFloat?
-    private var isDragging = false
-    private var draggingView: NSImageView?
-    private var draggingViewLeadingConstraint: NSLayoutConstraint?
-
-    open var favicon: NSImage? {
-        get {
-            return _favicon
-        }
-        set(image) {
-            _favicon = image as NSImage?
-            closeButton.image = _favicon
-        }
-    }
 
     /// The Tab Title
     open var label: String {
@@ -57,25 +37,26 @@ open class MATabBarItem: NSButton {
 
     override open func draw(_ dirtyRect: NSRect) {
         super.draw(dirtyRect)
-        // Tab Title
-        setUpTitle()
 
-        // Close Button
-        setUpCloseButton()
-
-        // Background and corner radius
-        let bgColor: NSColor = .quaternaryLabelColor
+        var bgColor: NSColor = configuration.lightTabColor
+        layer?.borderColor = configuration.lightTabBorderColor.cgColor
         layer?.cornerRadius = 4
+        layer?.borderWidth = 1
+        let appearance = UserDefaults.standard.string(forKey: "AppleInterfaceStyle") ?? "Light"
+
+        if appearance == "Dark" {
+            bgColor = configuration.darkTabColor
+            layer?.borderColor = configuration.darkTabBorderColor.cgColor
+        }
+
         layer?.masksToBounds = true
         layer?.backgroundColor = bgColor.cgColor
         bgColor.setFill()
         dirtyRect.fill()
-
-        let area = NSTrackingArea(rect: bounds, options: [.mouseEnteredAndExited, .activeAlways, .inVisibleRect], owner: self, userInfo: nil)
-        addTrackingArea(area)
     }
 
-    fileprivate func setUpTitle() {
+    private final func configureViews() {
+        // Tab Title
         let position = xPosition * 2 + closeButtonSize.width
 
         tabTitle.translatesAutoresizingMaskIntoConstraints = false
@@ -94,9 +75,8 @@ open class MATabBarItem: NSButton {
         tabTitle.setContentHuggingPriority(NSLayoutConstraint.Priority(rawValue: NSLayoutConstraint.Priority.defaultLow.rawValue - 10), for: .horizontal)
         tabTitle.setContentCompressionResistancePriority(NSLayoutConstraint.Priority.defaultLow, for: .horizontal)
         tabTitle.lineBreakMode = .byTruncatingTail
-    }
 
-    fileprivate func setUpCloseButton() {
+        // Close button
         closeButton.translatesAutoresizingMaskIntoConstraints = false
         closeButton.target = self
         closeButton.action = #selector(closeTab)
@@ -114,7 +94,12 @@ open class MATabBarItem: NSButton {
         closeButton.imagePosition = .imageOnly
         closeButton.layer?.masksToBounds = false
 
-        closeButton.image = favicon
+        tabTitle.stringValue = tab.title
+
+        let area = NSTrackingArea(rect: bounds, options: [.mouseEnteredAndExited, .activeAlways, .inVisibleRect], owner: self, userInfo: nil)
+        addTrackingArea(area)
+
+        closeButton.image = tab.icon
     }
 
     override open func mouseEntered(with event: NSEvent) {
@@ -125,13 +110,23 @@ open class MATabBarItem: NSButton {
 
     override open func mouseExited(with event: NSEvent) {
         super.mouseExited(with: event)
-        closeButton.image = favicon
-        animator().alphaValue = isMainButton ? 1 : 0.6
+        closeButton.image = tab.icon
+        animator().alphaValue = isSelectedTab ? 1 : 0.6
     }
 
     @objc func closeTab() {
-        delegate?.tabBarViewItem?(self, wantsToClose: tag)
+        delegate?.tabBarItem?(self, wantsToClose: tab)
     }
 
-    // TODO: - Dragging Session
+    public required init(frame frameRect: NSRect, tab: MATab) {
+        self.tab = tab
+        isSelectedTab = tab.isSelectedTab
+        super.init(frame: frameRect)
+        configureViews()
+    }
+
+    @available(*, unavailable)
+    public required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
 }