+++ Frameworks/MATabView/MATabView/MATabView.swift
@@ -8,142 +8,168 @@
 
 import Cocoa
 
-// MARK: - Delegate Methods
-
 @objc public protocol MATabViewDelegate: NSObjectProtocol {
     // Select Functions
-    @objc optional func tabView(_ tabView: MATabView, willSelect tabViewItemIndex: Int)
-    @objc optional func tabView(_ tabView: MATabView, didSelect tabViewItemIndex: Int)
+    @objc optional func tabView(_ tabView: MATabView, willSelect tab: MATab)
+    @objc optional func tabView(_ tabView: MATabView, didSelect tab: MATab)
 
     // Close Functions
-    @objc optional func tabView(_ tabView: MATabView, willClose tabViewItemIndex: Int)
-    @objc optional func tabView(_ tabView: MATabView, didClose tabViewItemIndex: Int)
+    @objc optional func tabView(_ tabView: MATabView, willClose tab: MATab)
+    @objc optional func tabView(_ tabView: MATabView, didClose tab: MATab)
 
     // Create New Tab Function
-    @objc optional func tabView(_ tabView: MATabView, willCreateTab tabViewItemIndex: Int)
-    @objc optional func tabView(_ tabView: MATabView, didCreateTab tabViewItemIndex: Int)
+    @objc optional func tabView(_ tabView: MATabView, willCreateTab tab: MATab)
+    @objc optional func tabView(_ tabView: MATabView, didCreateTab tab: MATab)
 
     // When there are no more tabs left
     @objc optional func tabView(noMoreTabsLeft tabView: MATabView)
 }
 
-open class MATabView: NSView, MATabBarViewDelegate {
-    open var selectedTabViewItemIndex: Int = 0
-    private var tabViews: [MATabViewItem] = []
+open class MATabView: NSView, MATabBarDelegate {
+    open var selectedTab: MATab?
+    open var tabs = [MATab]()
     open weak var delegate: MATabViewDelegate?
-    open var tabBar: MATabBarView?
+    private var tabBar = MATabBar(frame: .zero)
 
-    open var tabViewItems: Int {
-        return tabViews.count
+    open var configuration = MATabViewConfiguration()
+
+    // MARK: - Configuring
+
+    private final func configureViews() {
+        tabBar.delegate = self
+        addSubview(tabBar)
+
+        tabBar.translatesAutoresizingMaskIntoConstraints = false
+
+        NSLayoutConstraint.activate([
+            tabBar.heightAnchor.constraint(equalToConstant: configuration.tabHeight),
+            tabBar.topAnchor.constraint(equalTo: topAnchor),
+            tabBar.leadingAnchor.constraint(equalTo: leadingAnchor),
+            tabBar.trailingAnchor.constraint(equalTo: trailingAnchor),
+        ])
     }
 
-    override open func draw(_ dirtyRect: NSRect) {
+    override public func draw(_ dirtyRect: NSRect) {
         super.draw(dirtyRect)
-        tabBar?.delegate = self
+        let bgColor: NSColor = .white
+        layer?.masksToBounds = true
+        layer?.backgroundColor = bgColor.cgColor
+        bgColor.setFill()
+        dirtyRect.fill()
     }
 
-    open func getView(at index: Int) -> NSView? {
-        if index > tabViews.count {
-            return nil
-        }
-        return tabViews[index].view
+    public required init?(coder: NSCoder) {
+        super.init(coder: coder)
+        configureViews()
     }
 
-    public func tabBarView(_ tabBarView: MATabBarView, didSelect tabBarViewItemIndex: Int) {
-        selectTabViewItem(at: tabBarViewItemIndex)
+    override public required init(frame frameRect: NSRect) {
+        super.init(frame: frameRect)
+        configureViews()
     }
 
-    public func tabBarView(_ tabBarView: MATabBarView, wantsToClose tabBarViewItemIndex: Int) {
-        removeTabViewItem(at: tabBarViewItemIndex)
+    // MARK: - Tabs Controls
+
+    open func select(tab: Int) {
+        select(tab: tabs[tab])
     }
 
-    open func selectTabViewItem(at index: Int) {
-        selectTabViewItem(at: index, isNewTab: false)
-    }
+    open func select(tab: MATab, isNewTab: Bool = false) {
+        delegate?.tabView?(self, willSelect: tab)
 
-    private func selectTabViewItem(at index: Int, isNewTab: Bool = false) {
-        // Call the delegate function
-        delegate?.tabView?(self, willSelect: index)
-
-        if tabBar!.tabViewItems != 0 {
-            if let button = tabBar!.getTabItem(at: selectedTabViewItemIndex) {
-                button.animator().alphaValue = 0.6
-                button.isMainButton = false
-            }
-        }
-
-        for subview in subviews {
-            subview.removeFromSuperview()
-        }
-        // Add the new subview
-        let newView = tabViews[index].view!
-        autoresizesSubviews = true
-        newView.setFrameSize(frame.size)
-        addSubview(newView)
-        newView.autoresizingMask = [.width, .height]
-
-        selectedTabViewItemIndex = index
-
-        if tabBar!.tabViewItems != 0 {
-            if let button = tabBar!.getTabItem(at: selectedTabViewItemIndex) {
-                if !isNewTab {
-                    button.animator().alphaValue = 1
-                }
-                button.isMainButton = true
-            }
+        if !(tabBar.isEmpty()) {
+            let oldTab = tabBar.get(tabItem: selectedTab!.position)
+            oldTab.animator().alphaValue = 0.6
+            oldTab.isSelectedTab = false
         }
 
-        // Call the delegate function
-        delegate?.tabView?(self, didSelect: index)
-    }
+        selectedTab?.isSelectedTab = false
 
-    open func addTabViewItem(tabViewItem: MATabViewItem) {
-        delegate?.tabView?(self, willCreateTab: tabViews.count - 1)
+        selectedTab?.view.removeFromSuperview()
 
-        tabViews.append(tabViewItem)
+        let frameSize = NSSize(width: frame.width, height: frame.height - 30)
 
-        tabBar?.addTab(title: "New Tab")
-        selectTabViewItem(at: tabViews.count - 1, isNewTab: true)
-        if let button = tabBar!.getTabItem(at: selectedTabViewItemIndex) {
-            button.isMainButton = true
-        }
-        delegate?.tabView?(self, didCreateTab: tabViews.count - 1)
-    }
+        tab.view.setFrameSize(frameSize)
+        addSubview(tab.view)
+        tab.view.autoresizingMask = [.width, .height]
 
-    open func removeTabViewItem(at index: Int) {
-        delegate?.tabView?(self, willClose: index)
-        if selectedTabViewItemIndex == index {
-            // If there are no more tabs left
-            if index == 0, tabViews.count == 1 {
-                delegate?.tabView?(self, didClose: index)
-                delegate?.tabView?(noMoreTabsLeft: self)
-                return
-            } else if index == 0, tabViews.count != 1 {
-                selectTabViewItem(at: selectedTabViewItemIndex + 1)
-                selectedTabViewItemIndex -= 1
-                delegate?.tabView?(self, didClose: index)
-                tabBar?.removeTab(at: index)
-                tabViews.remove(at: index)
-                if let button = tabBar!.getTabItem(at: selectedTabViewItemIndex) {
-                    button.animator().alphaValue = 1
-                    button.isMainButton = true
-                }
-                return
-            } else {
-                selectTabViewItem(at: selectedTabViewItemIndex - 1)
-                delegate?.tabView?(self, didClose: index)
-                tabBar?.removeTab(at: index)
-                tabViews.remove(at: index)
-                if let button = tabBar!.getTabItem(at: selectedTabViewItemIndex) {
-                    button.animator().alphaValue = 1
-                    button.isMainButton = true
-                }
-                return
-            }
+        selectedTab = tab
+//
+        if !isNewTab {
+            let newTab = tabBar.get(tabItem: tab.position)
+            newTab.alphaValue = 0.0
+            newTab.isSelectedTab = true
+
+            newTab.animator().alphaValue = 1.0
         }
 
-        delegate?.tabView?(self, didClose: index)
-        tabBar?.removeTab(at: index)
-        tabViews.remove(at: index)
+        selectedTab!.isSelectedTab = true
+
+        delegate?.tabView?(self, didSelect: tab)
+    }
+
+    open func create(tab: MATab) {
+        delegate?.tabView?(self, willCreateTab: tab)
+
+        // Configure the tab
+        tab.position = tabs.count
+
+        tabs.append(tab)
+        select(tab: tab, isNewTab: true)
+        tabBar.add(tab: tab)
+
+        delegate?.tabView?(self, didCreateTab: tab)
+    }
+
+    open func remove(tab: Int) {
+        remove(tab: tabs[tab])
+    }
+
+    open func remove(tab: MATab) {
+        delegate?.tabView?(self, willClose: tab)
+
+        tabs.remove(at: tab.position)
+        tabBar.remove(tab: tab)
+
+        for (index, tab) in tabs.enumerated() {
+            tab.position = index
+        }
+
+        if !tabs.isEmpty {
+            selectedTab = selectedTab!.position == 0 ? tabs[selectedTab!.position + 1] : tabs[selectedTab!.position - 1]
+            select(tab: selectedTab!)
+
+            delegate?.tabView?(self, didClose: tab)
+        } else {
+            delegate?.tabView?(noMoreTabsLeft: self)
+        }
+    }
+
+    // MARK: - Tab Bar
+
+    public func tabBar(_ tabBarView: MATabBar, didSelect tab: MATab) {
+        select(tab: tab)
+    }
+
+    public func tabBar(_ tabBarView: MATabBar, wantsToClose tab: MATab) {
+        remove(tab: tab)
+    }
+
+    // MARK: - Get & Set
+
+    public func get(tab at: Int) -> MATab {
+        tabs[at]
+    }
+
+    open func set(tab at: Int, title: String) {
+        tabs[at].title = title
+        tabBar.get(tabItem: at).tab.title = title
+        tabBar.get(tabItem: at).tabTitle.stringValue = title
+    }
+
+    open func set(tab at: Int, icon: NSImage) {
+        tabs[at].icon = icon
+        tabBar.get(tabItem: at).tab.icon = icon
+        tabBar.get(tabItem: at).closeButton.image = icon
     }
 }