+++ Frameworks/MATabView/MATabView/Changes.md
@@ -1,6 +1,12 @@
 #  Changes
 
 ## v1.0
+- Added an example application
+- Made the code more simple
+- Added configurations
+- Fixed the scrolling issue
+- Re-wrote the entire tabing system
+- Changed the color of the tab bar items based on appearance
 - Fixed the bug where closing the tab may cause the application to crash
 - Added animations for when creating a new tab
 - Created more functions for getting the views